requirements.txt
torch==1.13.1
torchvision==0.14.1
numpy==1.26.4
pandas==2.2.2
pillow==10.4.0
fastapi==0.112.2
uvicorn==0.30.6
streamlit==1.38.0
opencv-python==4.10.0.84
albumentations==1.4.14
scikit-learn==1.5.1
matplotlib==3.9.2
seaborn==0.13.2
requests==2.32.3